chore: move config helper functions to utils, prep for adding tests
